refactor: factor out deserialize rule
