+++ crates/config/src/rule_config.rs
@@ -1,17 +1,13 @@
-use crate::referent_rule::{ReferentRule, ReferentRuleError, RuleRegistration};
-use crate::relational_rule::{Follows, Has, Inside, Precedes};
-use crate::rule::{
-  AtomicRule, CompositeRule, PatternStyle, RelationalRule, Rule, SerializableRule,
-};
+use crate::deserialize_env::DeserializeEnv;
+use crate::referent_rule::RuleRegistration;
+use crate::rule::{deserialize_rule, Rule, RuleSerializeError, SerializableRule};
 
 pub use crate::constraints::{
   try_deserialize_matchers, try_from_serializable as deserialize_meta_var, RuleWithConstraint,
   SerializableMetaVarMatcher, SerializeConstraintsError,
 };
 use ast_grep_core::language::Language;
-use ast_grep_core::matcher::{KindMatcher, KindMatcherError, RegexMatcher, RegexMatcherError};
 use ast_grep_core::meta_var::MetaVarMatchers;
-use ast_grep_core::ops as o;
 use ast_grep_core::replace_meta_var_in_string;
 use ast_grep_core::NodeMatch;
 use ast_grep_core::{Pattern, PatternError};
@@ -167,150 +163,6 @@ impl<L: Language> Deref for RuleConfig<L> {
   fn deref(&self) -> &Self::Target {
     &self.inner
   }
-}
-
-#[derive(Debug, Error)]
-pub enum RuleSerializeError {
-  #[error("Rule must have one positive matcher.")]
-  MissPositiveMatcher,
-  #[error("Rule contains invalid kind matcher.")]
-  InvalidKind(#[from] KindMatcherError),
-  #[error("Rule contains invalid pattern matcher.")]
-  InvalidPattern(#[from] PatternError),
-  #[error("Rule contains invalid regex matcher.")]
-  WrongRegex(#[from] RegexMatcherError),
-  #[error("Rule contains invalid matches reference.")]
-  MatchesRefrence(#[from] ReferentRuleError),
-  #[error("field is only supported in has/inside.")]
-  FieldNotSupported,
-}
-
-pub struct DeserializeEnv<L: Language> {
-  registration: RuleRegistration<L>,
-  lang: L,
-}
-
-impl<L: Language> DeserializeEnv<L> {
-  pub fn new(lang: L) -> Self {
-    Self {
-      registration: Default::default(),
-      lang,
-    }
-  }
-
-  pub fn register_utils(mut self, utils: &HashMap<String, SerializableRule>) -> RResult<Self> {
-    let registration = RuleRegistration::default();
-    for (id, rule) in utils {
-      let rule = RuleWithConstraint::new(
-        deserialize_rule(rule.clone(), &self)?,
-        MetaVarMatchers::default(),
-      );
-      if let Err(e) = registration.insert_rule(id, rule) {
-        return Err(RuleSerializeError::MatchesRefrence(e).into());
-      }
-    }
-    self.registration = registration;
-    Ok(self)
-  }
-  pub fn with_registration(mut self, registration: RuleRegistration<L>) -> Self {
-    self.registration = registration;
-    self
-  }
-}
-
-// TODO: implement positive/non positive
-pub fn deserialize_rule<L: Language>(
-  serialized: SerializableRule,
-  env: &DeserializeEnv<L>,
-) -> Result<Rule<L>, RuleSerializeError> {
-  let mut rules = Vec::with_capacity(1);
-  use Rule as R;
-  let categorized = serialized.categorized();
-  deserialze_atomic_rule(categorized.atomic, &mut rules, env)?;
-  deserialize_relational_rule(categorized.relational, &mut rules, env)?;
-  deserialze_composite_rule(categorized.composite, &mut rules, env)?;
-  if rules.is_empty() {
-    Err(RuleSerializeError::MissPositiveMatcher)
-  } else if rules.len() == 1 {
-    Ok(rules.pop().expect("should not be empty"))
-  } else {
-    Ok(R::All(o::All::new(rules)))
-  }
-}
-
-fn deserialze_composite_rule<L: Language>(
-  composite: CompositeRule,
-  rules: &mut Vec<Rule<L>>,
-  env: &DeserializeEnv<L>,
-) -> Result<(), RuleSerializeError> {
-  use Rule as R;
-  let convert_rules = |rules: Vec<SerializableRule>| -> Result<_, RuleSerializeError> {
-    let mut inner = Vec::with_capacity(rules.len());
-    for rule in rules {
-      inner.push(deserialize_rule(rule, env)?);
-    }
-    Ok(inner)
-  };
-  if let Some(all) = composite.all {
-    rules.push(R::All(o::All::new(convert_rules(all)?)));
-  }
-  if let Some(any) = composite.any {
-    rules.push(R::Any(o::Any::new(convert_rules(any)?)));
-  }
-  if let Some(not) = composite.not {
-    let not = o::Not::new(deserialize_rule(*not, env)?);
-    rules.push(R::Not(Box::new(not)));
-  }
-  if let Some(id) = composite.matches {
-    let matches = ReferentRule::try_new(id, &env.registration)?;
-    rules.push(R::Matches(matches));
-  }
-  Ok(())
-}
-
-fn deserialize_relational_rule<L: Language>(
-  relational: RelationalRule,
-  rules: &mut Vec<Rule<L>>,
-  env: &DeserializeEnv<L>,
-) -> Result<(), RuleSerializeError> {
-  use Rule as R;
-  // relational
-  if let Some(inside) = relational.inside {
-    rules.push(R::Inside(Box::new(Inside::try_new(*inside, env)?)));
-  }
-  if let Some(has) = relational.has {
-    rules.push(R::Has(Box::new(Has::try_new(*has, env)?)));
-  }
-  if let Some(precedes) = relational.precedes {
-    rules.push(R::Precedes(Box::new(Precedes::try_new(*precedes, env)?)));
-  }
-  if let Some(follows) = relational.follows {
-    rules.push(R::Follows(Box::new(Follows::try_new(*follows, env)?)));
-  }
-  Ok(())
-}
-
-fn deserialze_atomic_rule<L: Language>(
-  atomic: AtomicRule,
-  rules: &mut Vec<Rule<L>>,
-  env: &DeserializeEnv<L>,
-) -> Result<(), RuleSerializeError> {
-  use Rule as R;
-  if let Some(pattern) = atomic.pattern {
-    rules.push(match pattern {
-      PatternStyle::Str(pat) => R::Pattern(Pattern::try_new(&pat, env.lang.clone())?),
-      PatternStyle::Contextual { context, selector } => {
-        R::Pattern(Pattern::contextual(&context, &selector, env.lang.clone())?)
-      }
-    });
-  }
-  if let Some(kind) = atomic.kind {
-    rules.push(R::Kind(KindMatcher::try_new(&kind, env.lang.clone())?));
-  }
-  if let Some(regex) = atomic.regex {
-    rules.push(R::Regex(RegexMatcher::try_new(&regex)?));
-  }
-  Ok(())
 }
 
 #[cfg(test)]
